+++ packages/react-mui-audit-log-viewer/src/components/AuditLogViewerComponent/index.tsx
@@ -72,6 +72,8 @@ const AuditLogViewerComponent: FC<ViewerProps> = ({
   searchOnChange = true,
   searchOnFilterChange = true,
   searchOnMount = true,
+
+  initialQuery = "",
 }) => {
   const {
     visibilityModel,
@@ -96,7 +98,7 @@ const AuditLogViewerComponent: FC<ViewerProps> = ({
   const filterFields = useAuditFilterFields(schema);
   const conditionalOptions = useAuditConditionalOptions(schema);
 
-  const hasMountedRef = useRef(false);
+  const hasMountedRef = useRef<string | undefined>(undefined);
 
   const bodyRef = useRef(body);
   bodyRef.current = body;
@@ -107,27 +109,34 @@ const AuditLogViewerComponent: FC<ViewerProps> = ({
   };
 
   useEffect(() => {
+    const initialQuery_ = initialQuery ?? "";
     if (
       !!bodyWithoutQuery &&
       (searchOnFilterChange || searchOnMount) &&
       !searchOnChange
     ) {
-      if (!searchOnFilterChange && hasMountedRef.current === true) {
+      if (!searchOnFilterChange && hasMountedRef.current === initialQuery_) {
         return;
       }
 
-      if (!searchOnMount && hasMountedRef.current === false) {
-        hasMountedRef.current = true;
+      if (!searchOnMount && hasMountedRef.current === undefined) {
+        hasMountedRef.current = initialQuery_;
         return;
       }
 
-      hasMountedRef.current = true;
+      hasMountedRef.current = initialQuery_;
       setTimeout(() => {
         handleSearch();
         // Add slight delay since since filters may update the query string
       }, 100);
     }
-  }, [bodyWithoutQuery, searchOnFilterChange, searchOnMount, searchOnChange]);
+  }, [
+    bodyWithoutQuery,
+    searchOnFilterChange,
+    searchOnMount,
+    searchOnChange,
+    initialQuery,
+  ]);
 
   useEffect(() => {
     if (!!body && searchOnChange) handleSearch();