+++ packages/react-mui-audit-log-viewer/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2024-11-14
+
+### Fixed
+
+- Updated to "@pangeacyber/react-mui-shared": "1.0.2",
+- Updated searchOnMount to re-search if initialQuery changes
+- Updated to diplay seconds for time fields
+
 ## [1.0.3] - 2024-11-13
 
 ### Fixed